MOONRAY-5199 macos merge (#890)
* Apple source

* Internal build fixes

* Fix renamed modules

* Platform fixes

* Platform fixes

* Fix compile defs

* Platform fixes

* Fix rdl2_ispc_util build problem

* No cpu affinity for mac

* cmake_policy CMP0144

* Revert "cmake_policy CMP0144"

This reverts commit dc7b54dcd31d29f3a2d76253923ac63aaa523cbe.

* Apple update

* fix unit test working directory

* Apple fixes

* Fixes for unit tests, wip

+++ tests/lib/scene/rdl2/TestValueContainer.cc
@@ -744,7 +744,7 @@ TestValueContainer::testLongVector() // int64_t
                  return total;
              },
              [&](ValueContainerDeq *vcDeq) { // deqFunc
-                 std::vector<long> pVec = vcDeq->deqLongVector();
+                 std::vector<Long> pVec = vcDeq->deqLongVector();
                  CPPUNIT_ASSERT(compareVector(vec, pVec));
              });
 }
@@ -1209,12 +1209,12 @@ TestValueContainer::testVLLongVector() // int64_t
 
                  size_t total = ValueContainerUtil::variableLengthEncodingSize(vec.size());
                  for (const auto &itr: vec) {
-                     total += ValueContainerUtil::variableLengthEncodingSize(itr);
+                     total += ValueContainerUtil::variableLengthEncodingSize(long(itr));
                  }
                  return total;
              },
              [&](ValueContainerDeq *vcDeq) { // deqFunc
-                 std::vector<long> pVec = vcDeq->deqVLLongVector();
+                 std::vector<Long> pVec = vcDeq->deqVLLongVector();
                  CPPUNIT_ASSERT(compareVector(vec, pVec));
              });
 }
